Tidied up a few minor syntax errors that were breaking the build with Visual Studio and added a few casts to get rid of warnings
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@295444 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_msg_buff.c
@@ -191,7 +191,7 @@ void jk_b_end(jk_msg_buf_t *msg, int protoh)
      * Ugly way to set the size in the right position 
      */
     jk_b_set_int(msg, 2, (unsigned short )(msg->len - 4)); /* see protocol */
-    jk_b_set_int(msg, 0, protoh);
+    jk_b_set_int(msg, 0, (unsigned short) protoh);
 }
